% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPredictionUmap}
\alias{plotPredictionUmap}
\title{Plot scTME Predictions on a UMAP}
\usage{
plotPredictionUmap(seurat_obj, reduction = "umap", pred_col = "prediction")
}
\arguments{
\item{seurat_obj}{A Seurat object containing scTME predictions in the metadata.}

\item{reduction}{A character string specifying the dimensional reduction to use. Default is "umap".}

\item{pred_col}{A character string specifying the metadata column containing predictions.}
}
\description{
Plot scTME Predictions on a UMAP
}
