% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scTME_atlas}
\alias{scTME_atlas}
\title{scTME Reference Atlas}
\format{
A Seurat object.
}
\usage{
scTME_atlas
}
\description{
A 50,000 cell reference dataset for scTME tumor microenvironment predictions.
}
\keyword{datasets}
