% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{scTME}
\alias{scTME}
\title{Run scTME Predictions}
\usage{
scTME(
  obj,
  ref = NULL,
  clusters = NULL,
  level = 2,
  label.col = NULL,
  max.cells = 50000,
  fast = FALSE,
  return.details = FALSE
)
}
\arguments{
\item{obj}{The query dataset}

\item{ref}{The reference atlas}

\item{clusters}{Optional cluster labels}

\item{level}{Resolution level (1 or 2)}

\item{label.col}{Target metadata column}

\item{max.cells}{Downsampling threshold}

\item{fast}{Downsample the reference atlas}

\item{return.details}{Return prediction strength metrics}
}
\description{
Run scTME Predictions
}
